+++ buildiOS_CI.sh
@@ -1,16 +1,14 @@
 #!/bin/bash
 
-# set -euo pipefail
-# set -eo pipefail
-set -o pipefail
+set -euo pipefail
 set -x
 
 ## Openssl_curl done previously
-# cd openssl_curl_for_ios_android.20170105
+# cd openssl_curl
 
 # SDK Version
 SDK_VERSION="17.2"
-MIN_VERSION="10.0"
+MIN_VERSION="13.0"
 
 # Setup paths
 # echo current directory is $(pwd)
@@ -63,7 +61,12 @@ build_AWSRelease_bitcode()
     cd "$BUILD_FOLDER"
 
     # Cleanup
-    rm -r ./*
+    if [ -z "$( ls -A . )" ]; then
+        echo "$BUILD_FOLDER empty"
+    else
+        echo "Cleanup $BUILD_FOLDER"
+        rm -r ./*
+    fi
 
 	export SDK=$2
 
@@ -82,7 +85,7 @@ build_AWSRelease_bitcode()
 
     cmake -Wno-dev \
         -DCMAKE_OSX_SYSROOT="$IPHONEOS_SDK" \
-        -DCMAKE_OSX_ARCHITECTURES="$ARCH" \
+        -DCMAKE_OSX_ARCHITECTURES=$ARCH \
         -DCMAKE_SYSTEM_NAME="Darwin" \
 		-DENABLE_TESTING=0 \
 		-DBUILD_ONLY="kinesis;cognito-identity;cognito-sync;lambda;s3;apigateway;identity-management" \
@@ -99,13 +102,10 @@ build_AWSRelease_bitcode()
 		-DCMAKE_CXX_FLAGS=-O3 \
 		-DCPP_STANDARD=14 \
 		-DENABLE_CURL_CLIENT=Yes \
-		-DCURL_INCLUDE_DIR=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/curl-ios-$ARCH/include \
-		-DCURL_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/curl-ios-$ARCH/lib/libcurl.a \
-		-DCMAKE_IOS_DEPLOYMENT_TARGET="12" \
-		-DENABLE_OPENSSL_ENCRYPTION=Yes \
-		-DOPENSSL_CRYPTO_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/lib/libcrypto.a \
-		-DOPENSSL_SSL_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/lib/libssl.a \
-		-DOPENSSL_INCLUDE_DIR=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/include \
+		-DCURL_INCLUDE_DIR=${WORKSPACE}/openssl_curl/libcurl-8.6.0-openssl-3.0.13-nghttp2-1.60.0/include \
+		-DCURL_LIBRARY=${WORKSPACE}/openssl_curl/libcurl-8.6.0-openssl-3.0.13-nghttp2-1.60.0/lib/iOS/libcurl.a \
+		-DCMAKE_IOS_DEPLOYMENT_TARGET=“12” \
+		-DUSE_OPENSSL=OFF \
         $WORKSPACE/aws-sdk-cpp
     make -j 8
     make install
@@ -116,183 +116,6 @@ build_AWSRelease_bitcode()
 }
 
 
-# ----------------------------------
-# To build arm64, armv7, armv7s
-# ----------------------------------
-build_AWSDebug_bitcode()
-{
-	export ARCH=$1
-	export DBGREL="Debug"
-	export BUILD_FOLDER=build/iOS-$ARCH-$DBGREL
-
-    # Build intermediates dir
-    mkdir -p "$BUILD_FOLDER"
-    cd "$BUILD_FOLDER"
-
-    # Cleanup
-    rm -r ./*
-
-	export SDK=$2
-
-    export CC="$(xcrun -sdk iphoneos -find clang)"
-    export CPP="$CC -E"
-    export CFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION"
-    export AR=$(xcrun -sdk iphoneos -find ar)
-    export RANLIB=$(xcrun -sdk iphoneos -find ranlib)
-    export CPPFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION"
-    export LDFLAGS="-arch ${ARCH} -isysroot $SDK"
-
-    BUILD_OUTPUT="${WORKSPACE}/output/iOS/${ARCH}/${DBGREL}"
-    mkdir -p "$BUILD_OUTPUT"
-
-    echo "$CMAKE_CXX_FLAGS"
-
-    cmake -Wno-dev \
-        -DCMAKE_OSX_SYSROOT="$IPHONEOS_SDK" \
-        -DCMAKE_OSX_ARCHITECTURES="$ARCH" \
-        -DCMAKE_SYSTEM_NAME="Darwin" \
-		-DENABLE_TESTING=0 \
-		-DBUILD_ONLY="kinesis;cognito-identity;cognito-sync;lambda;s3;apigateway;identity-management" \
-        -DCMAKE_SHARED_LINKER_FLAGS="-framework Foundation -lz -framework Security" \
-        -DCMAKE_EXE_LINKER_FLAGS="-framework Foundation -framework Security" \
-        -DCMAKE_PREFIX_PATH="$WORKSPACE/libcurl/" \
-        -DBUILD_SHARED_LIBS=OFF \
-        -DCUSTOM_MEMORY_MANAGEMENT=0 \
-        -DCMAKE_BUILD_TYPE=$DBGREL \
-        -DCMAKE_INSTALL_PREFIX="$BUILD_OUTPUT" \
-        -DCMAKE_CXX_FLAGS="-std=c++11 -stdlib=libc++ -miphoneos-version-min=$MIN_VERSION" \
-        $WORKSPACE/aws-sdk-cpp
-    make -j 8
-    make install
-
-    # Go back
-    cd ..
-	cd ..
-}
-
-
-
-
-# ----------------------------------
-# To build i386, x86_64 (fails currently)
-# ----------------------------------
-build_AWSRelease_Simulator_bitcode()
-{
-    export ARCH=$1
-	export DBGREL="Release"
-	export BUILD_FOLDER=build/iOS-$ARCH-$DBGREL
-
-    # Build intermediates dir
-    mkdir -p "$BUILD_FOLDER"
-    cd "$BUILD_FOLDER"
-
-    # Cleanup
-    rm -r ./*
-
-    export SDK=$2
-
-    export CC="$(xcrun -sdk iphonesimulator -find clang)"
-    export CPP="$CC -E"
-    export CFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION"
-    export AR=$(xcrun -sdk iphonesimulator -find ar)
-    export RANLIB=$(xcrun -sdk iphonesimulator -find ranlib)
-    export CPPFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION -Wno-error"
-    export LDFLAGS="-arch ${ARCH} -isysroot $SDK"
-
-    BUILD_OUTPUT="${WORKSPACE}/output/iOS/${ARCH}/${DBGREL}"
-    mkdir -p "$BUILD_OUTPUT"
-
-    echo "$CMAKE_CXX_FLAGS"
-
-    cmake -Wno-dev \
-    -DCMAKE_OSX_SYSROOT=$IPHONESIMULATOR_SDK \
-    -DCMAKE_OSX_ARCHITECTURES="$ARCH" \
-    -DCMAKE_SYSTEM_NAME="Darwin" \
-	-DENABLE_TESTING=0 \
-	-DBUILD_ONLY="kinesis;cognito-identity;cognito-sync;lambda;s3;apigateway;identity-management" \
-    -DCMAKE_SHARED_LINKER_FLAGS="-framework Foundation -lz -framework Security" \
-    -DCMAKE_EXE_LINKER_FLAGS="-framework Foundation -framework Security" \
-    -DCMAKE_PREFIX_PATH="$WORKSPACE/libcurl/" \
-    -DBUILD_SHARED_LIBS=OFF \
-    -DCUSTOM_MEMORY_MANAGEMENT=0 \
-    -DCMAKE_BUILD_TYPE=$DBGREL \
-    -DCMAKE_INSTALL_PREFIX="$BUILD_OUTPUT" \
-    -DCMAKE_CXX_FLAGS="-std=c++11 -stdlib=libc++ -miphoneos-version-min=$MIN_VERSION" \
-	-DSTATIC_LINKING=1 \
-	-DTARGET_ARCH=APPLE \
-	-DCMAKE_CXX_FLAGS=-O3 \
-	-DCPP_STANDARD=14 \
-	-DENABLE_CURL_CLIENT=Yes \
-	-DCURL_INCLUDE_DIR=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/curl-ios-$ARCH/include \
-	-DCURL_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/curl-ios-$ARCH/lib/libcurl.a \
-	-DCMAKE_IOS_DEPLOYMENT_TARGET="12" \
-	-DENABLE_OPENSSL_ENCRYPTION=Yes \
-	-DOPENSSL_CRYPTO_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/lib/libcrypto.a \
-	-DOPENSSL_SSL_LIBRARY=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/lib/libssl.a \
-	-DOPENSSL_INCLUDE_DIR=${WORKSPACE}/openssl_curl_for_ios_android.20170105/output/ios/openssl-ios-$ARCH/include \
-    $WORKSPACE/aws-sdk-cpp
-    make -j 8
-    make install
-
-    # Go back
-    cd ..
-	cd ..
-}
-
-
-# ----------------------------------
-# To build i386, x86_64 (fails currently)
-# ----------------------------------
-build_AWSDebug_Simulator_bitcode()
-{
-    export ARCH=$1
-	export DBGREL="Debug"
-	export BUILD_FOLDER=build/iOS-$ARCH-$DBGREL
-
-    # Build intermediates dir
-    mkdir -p "$BUILD_FOLDER"
-    cd "$BUILD_FOLDER"
-
-    # Cleanup
-    rm -r ./*
-
-    export SDK=$2
-
-    export CC="$(xcrun -sdk iphonesimulator -find clang)"
-    export CPP="$CC -E"
-    export CFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION"
-    export AR=$(xcrun -sdk iphonesimulator -find ar)
-    export RANLIB=$(xcrun -sdk iphonesimulator -find ranlib)
-    export CPPFLAGS="-arch ${ARCH} -isysroot $SDK -miphoneos-version-min=$MIN_VERSION -Wno-error"
-    export LDFLAGS="-arch ${ARCH} -isysroot $SDK"
-
-    BUILD_OUTPUT="${WORKSPACE}/output/iOS/${ARCH}/${DBGREL}"
-    mkdir -p "$BUILD_OUTPUT"
-
-    echo "$CMAKE_CXX_FLAGS"
-
-    cmake -Wno-dev \
-    -DCMAKE_OSX_SYSROOT=$IPHONESIMULATOR_SDK \
-    -DCMAKE_OSX_ARCHITECTURES=$ARCH \
-    -DCMAKE_SYSTEM_NAME="Darwin" \
-	-DENABLE_TESTING=0 \
-	-DBUILD_ONLY="kinesis;cognito-identity;cognito-sync;lambda;s3;apigateway;identity-management" \
-    -DCMAKE_SHARED_LINKER_FLAGS="-framework Foundation -lz -framework Security" \
-    -DCMAKE_EXE_LINKER_FLAGS="-framework Foundation -framework Security" \
-    -DCMAKE_PREFIX_PATH="$WORKSPACE/libcurl/" \
-    -DBUILD_SHARED_LIBS=OFF \
-    -DCUSTOM_MEMORY_MANAGEMENT=0 \
-    -DCMAKE_BUILD_TYPE=$DBGREL \
-    -DCMAKE_INSTALL_PREFIX="$BUILD_OUTPUT" \
-    -DCMAKE_CXX_FLAGS="-std=c++11 -stdlib=libc++ -miphoneos-version-min=$MIN_VERSION" \
-    $WORKSPACE/aws-sdk-cpp
-    make -j 8
-    make install
-
-    # Go back
-    cd ..
-	cd ..
-}
 
 # ---------------------------------------------------
 # Function to aggregate build outputs into a fat lib
@@ -355,71 +178,13 @@ aggregate_release_libs() {
     echo "--------------------------------------------------"
 }
 
-# ---------------------------------------------------
-# Function to aggregate build outputs into a fat lib
-# ---------------------------------------------------
-aggregate_debug_libs() {
-    set +x
-    AGG_OUTPUT_DIR=$1
-	export DBGREL="Debug"
-
-    # Aggregate library and include files
-    mkdir -p ${AGG_OUTPUT_DIR}/include
-    mkdir -p ${AGG_OUTPUT_DIR}/lib
-
-    cp -r ${WORKSPACE}/output/iOS/arm64/${DBGREL}/include/* ${AGG_OUTPUT_DIR}/include/
-
-    ## declare an array variable with required aws components
-    ## This is an example. Change as required
-    declare -a components=( "access-management"
-                            "cognito-identity"
-                            "cognito-sync"
-							"core"
-                            "iam"
-                            "kinesis"
-                            "lambda"
-                            )
-
-    ## now loop through the above array
-    for component in "${components[@]}"
-    do
-        LIBNAME="libaws-cpp-sdk-${component}.a"
-        echo "--------- Aggregating $LIBNAME ---------"
-        xcrun -sdk iphoneos lipo \
-        "${WORKSPACE}/output/iOS/arm64/${DBGREL}/lib/${LIBNAME}" \
-        "${WORKSPACE}/output/iOS/armv7/${DBGREL}/lib/${LIBNAME}" \
-        "${WORKSPACE}/output/iOS/x86_64/${DBGREL}/lib/${LIBNAME}" \
-        -create -output ${AGG_OUTPUT_DIR}/lib/${LIBNAME}
-
-        # verify arch
-        echo "- Running lipo info for $LIBNAME:"
-        lipo -info ${AGG_OUTPUT_DIR}/lib/${LIBNAME}
-        echo "--------------------------------------------------"
-    done
-
-    echo "--------------------------------------------------"
-    echo "Aggregated output location: ${AGG_OUTPUT_DIR}"
-    echo "--------------------------------------------------"
-}
 
 
 ## Build release configuration
 build_AWSRelease_bitcode "arm64" "${IPHONEOS_SDK}"
-build_AWSRelease_bitcode "armv7" "${IPHONEOS_SDK}"
-#build_AWSRelease_bitcode "armv7s" "${IPHONEOS_SDK}"
-#build_AWSRelease_Simulator_bitcode "i386" "${IPHONESIMULATOR_SDK}"
-build_AWSRelease_Simulator_bitcode "x86_64" "${IPHONESIMULATOR_SDK}"
-
-## Build debug configuration
-#build_AWSDebug_bitcode "arm64" "${IPHONEOS_SDK}"
-#build_AWSDebug_bitcode "armv7" "${IPHONEOS_SDK}"
-#build_AWSDebug_bitcode "armv7s" "${IPHONEOS_SDK}"
-#build_AWSDebug_Simulator_bitcode "i386" "${IPHONESIMULATOR_SDK}"
-#build_AWSDebug_Simulator_bitcode "x86_64" "${IPHONESIMULATOR_SDK}"
 
 ## Aggregate into a fat lib. Argument provided here is the output directory
-aggregate_release_libs "${WORKSPACE}/output/iOS/fatlib/release"
-#aggregate_debug_libs "${WORKSPACE}/output/iOS/fatlib/debug"
+#aggregate_release_libs "${WORKSPACE}/output/iOS/fatlib/release"
 
 echo "-- Initial AWSSDKCPP build complete --"
